ml_models_train/ner_models/address_recognition_flair/requirements.txt
flair
transformers
scikit-learn
pandas
numpy
boto3
python-dotenv